Move all stream wrapper utils into one place
